requirements.txt
python-dotenv
python-telegram-bot
requests
beautifulsoup4
pandas
python-jobspy
firebase-admin
dateparser